+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+### Fixed
+- added terminals list file check
+
 ## [1.9.1]
 ### Fixed
 - fixed PHP errors, when new settings is not saved